pyproject.toml
[project]
name = "genesis"
version = "0.1.0"
description = "Brain-dump to execution pipeline — Google ADK + MCP + AlloyDB"
requires-python = ">=3.12,<3.13"
dependencies = [
    # Google ADK + Vertex AI
    "google-adk==1.28.1",
    "google-genai==1.70.0",
    "google-cloud-aiplatform==1.145.0",

    # FastAPI
    "fastapi==0.135.3",
    "uvicorn[standard]==0.44.0",
    "sse-starlette==3.3.4",

    # Pydantic
    "pydantic==2.12.5",

    # AlloyDB / Postgres
    "asyncpg==0.31.0",
    "sqlalchemy[asyncio]==2.0.49",
    "google-cloud-alloydb-connector[asyncpg]==1.12.1",

    # Google API clients (Docs, Calendar)
    "google-api-python-client==2.193.0",
    "google-auth==2.49.1",
    "google-auth-httplib2==0.3.1",

    # MCP Toolbox
    "toolbox-core==1.0.0",

    # Utilities
    "python-dotenv==1.2.2",
    "httpx==0.28.1",
    "tenacity==9.1.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "api", "db", "mcp"]

[tool.uv]
dev-dependencies = []
